+++ previous-versions/assess/src/codatassess/models/shared/__init__.py
@@ -5,7 +5,6 @@ from .accountcategoriesupdatedwebhookdata import *
 from .accountcategorylevel import *
 from .accountref import *
 from .accounts import *
-from .bankingtransactionref import *
 from .commercereportcomponent import *
 from .commercereportdimension import *
 from .commercereporterror import *
@@ -26,7 +25,6 @@ from .details import *
 from .enhancedcashflowitem import *
 from .enhancedcashflowtransaction import *
 from .enhancedcashflowtransactions import *
-from .enhancedinvoicereportitem import *
 from .enhancedinvoicesreport import *
 from .enhancedreport import *
 from .enhancedreportaccountcategory import *
@@ -37,9 +35,7 @@ from .excelreporttypes import *
 from .excelstatus import *
 from .halref import *
 from .integritystatus import *
-from .invoicestatus import *
 from .itemref import *
-from .lendingcustomerref import *
 from .links import *
 from .loanref import *
 from .loansummary import *
@@ -50,7 +46,6 @@ from .loansummaryreportinfo import *
 from .loansummaryreportitem import *
 from .loantransactions import *
 from .loantransactionsreportinfo import *
-from .payment import *
 from .periodunit import *
 from .report import *
 from .reportcomponentmeasure import *
@@ -62,4 +57,4 @@ from .status import *
 from .summaries import *
 from .transactioncategory import *
 
-__all__ = ["AccountCategoriesUpdatedWebhook","AccountCategoriesUpdatedWebhookData","AccountCategoryLevel","AccountRef","Accounts","BankingTransactionRef","CommerceReportComponent","CommerceReportDimension","CommerceReportDimensionItems","CommerceReportError","CommerceReportMeasure","DataIntegrityAmounts","DataIntegrityByAmount","DataIntegrityByCount","DataIntegrityConnectionID","DataIntegrityDataType","DataIntegrityDates","DataIntegrityDetails","DataIntegrityMatch","DataIntegrityStatus","DataIntegrityStatusInfo","DataIntegritySummary","DataSource","Details","EnhancedCashFlowItem","EnhancedCashFlowTransaction","EnhancedCashFlowTransactions","EnhancedInvoiceReportItem","EnhancedInvoicesReport","EnhancedReport","EnhancedReportAccountCategory","EnhancedReportInfo","EnhancedReportReportItem","ErrorMessage","ExcelReportType","ExcelReportTypes","ExcelStatus","HalRef","IntegrityStatus","InvoiceStatus","ItemRef","LendingCustomerRef","Links","LoanRef","LoanSummary","LoanSummaryIntegrationType","LoanSummaryRecordRef","LoanSummaryRecordRefType","LoanSummaryReportInfo","LoanSummaryReportItem","LoanTransactions","LoanTransactionsReportInfo","Payment","PeriodUnit","Report","ReportComponentMeasure","ReportInfo","ReportItems","ReportItemsLoanTransactionType","Security","SourceRef","Status","Summaries","TransactionCategory"]
+__all__ = ["AccountCategoriesUpdatedWebhook","AccountCategoriesUpdatedWebhookData","AccountCategoryLevel","AccountRef","Accounts","CommerceReportComponent","CommerceReportDimension","CommerceReportDimensionItems","CommerceReportError","CommerceReportMeasure","DataIntegrityAmounts","DataIntegrityByAmount","DataIntegrityByCount","DataIntegrityConnectionID","DataIntegrityDataType","DataIntegrityDates","DataIntegrityDetails","DataIntegrityMatch","DataIntegrityStatus","DataIntegrityStatusInfo","DataIntegritySummary","DataSource","Details","EnhancedCashFlowItem","EnhancedCashFlowTransaction","EnhancedCashFlowTransactions","EnhancedInvoicesReport","EnhancedReport","EnhancedReportAccountCategory","EnhancedReportInfo","EnhancedReportReportItem","ErrorMessage","ExcelReportType","ExcelReportTypes","ExcelStatus","HalRef","IntegrityStatus","ItemRef","Links","LoanRef","LoanSummary","LoanSummaryIntegrationType","LoanSummaryRecordRef","LoanSummaryRecordRefType","LoanSummaryReportInfo","LoanSummaryReportItem","LoanTransactions","LoanTransactionsReportInfo","PeriodUnit","Report","ReportComponentMeasure","ReportInfo","ReportItems","ReportItemsLoanTransactionType","Security","SourceRef","Status","Summaries","TransactionCategory"]