+++ previous-versions/assess/src/codatassess/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0'
-    sdk_version: str = '0.33.0'
-    gen_version: str = '2.139.1'
-    user_agent: str = 'speakeasy-sdk/python 0.33.0 2.139.1 1.0 codat-assess'
+    sdk_version: str = '0.34.0'
+    gen_version: str = '2.142.2'
+    user_agent: str = 'speakeasy-sdk/python 0.34.0 2.142.2 1.0 codat-assess'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: