+++ previous-versions/assess/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.33.0] previous-versions/assess
 ### Releases
 - [PyPI v0.33.0] https://pypi.org/project/codat-assess/0.33.0 - previous-versions/assess
+
+## 2023-10-01 00:14:22
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.92.2 (2.142.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] previous-versions/assess
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-assess/0.34.0 - previous-versions/assess